Allow installApk to install a testOnly APK
Just makes it a little more resilient, in the sense that if you built the APK from the IDE and haven't set testOnly="false" in the manifest, this will still successfully install it, which one would probably expect for an adb-commanding tool. Just ran into this with composer and found the install command here.